Move admin and model to separate files
