+++ core/models/__init__.py
@@ -16,3 +16,4 @@ from .substance import *
 from .time_frame import *
 from .usage import *
 from .user import *
+from .annual_project_report import *
